Temporarily exclude libpathTestRtf and libpathTestRtfChild tests on OSX
libpathTestRtf and libpathTestRtfChild tests pass locally on OSX. But,
they hang in Jenkins builds due to a pop-up requiring user-action. This
causes a Jenkins build to halt for 4 hrs. Temporarily excluding
libpathTestRtf and libpathTestRtfChild tests on OSX to avoid the halt
for 4 hrs. The tests will be added back once the issue is resolved.

Issue: https://github.com/eclipse/openj9/issues/3787

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>