Don't build JitBuilderRecorder files from OMR yet
An OMR pull request will be merged soon that introduces four new
files as part of the JitBuilder library. While I am in the process
of bringing JitBuilder support into OpenJ9, I am not sure if these
files will compile or not as part of OpenJ9, so let's leave them
out for now. They are useless to OpenJ9 now anyway.

I'll bring these files back in as part of the larger effort
to bring JitBuilder into OpenJ9.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>