oFree client specific data on VM shutdown
JITaaS server caches the information for the JITaaS client
in an internal hash map per client.This data is purged if
not used for sometime(1000 minutes) but this can grow very
large if the number of client increases to large number.
To fix this issue the client data needs to be deleted when
the client compilation thread teminates.

On VM shutdown, stopCompilationThreads get called which terminate
the client thread.On thread temination a dummy remote compilation
request is build with NULL J9Method. On getting this request the server
check if the inUse for the client session data is 0 and removes the client
session data from the hash map. If the inUse is not 0 then the
the client session data is marked to be deleted and the onus of deleting
the client session data lies with whoever make the inUse 0.

Signed-off-by: Satbir Singh <satbir.singh1@ibm.com>