Require interface in resolveInterfaceMethodRefInto
This function should fail whenever interfaceClass is not an interface,
even if no exception is to be thrown. Currently it can continue on with
a non-interface class and find a method that belongs neither to an
interface nor to Object, which is supposed to be impossible for
interface method references.

Fixes #3916
Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>