Increase the number of compilation threads on the JITaaS server
Currently the JIT allows a maximum of 8 compilation threads due to
the size of uint8_t TR_PersistentClassInfo::_shouldNotBeNewlyExtended.
The JITaaS server uses a different mechanism to track newly extended
classes and it does not set the bit in _shouldNotBeNewlyExtended.
This change increases the max number of compilation threads
on the JIT server to 64.

Closes: #3526

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>