Remove inconsistant DDR !dumppackage test
The package `com/ibm/gpu/spi` is not always exported to all unnamed
modules. As a result, one of the DDR !dumppackage tests was
inconsistent. This lack of being exported seemed to occur on windows
machines.

The requirement for said package to be exported to all unnamed
modules has been removed as this was not a critical part of the test.

This solves #4028.

Signed-off-by: Andrew Crowther <acrowthe3388@gmail.com>