Add value type static preloading and tests
Implement the following upon preloading a valueType:
- Preload static fields with a Q signature
- Throw an `IncompatibleClassChangeError` if the loaded class is not a
value type
- Throw a `NoClassDefError` if the class is not defined

In addition, tests were added for these cases. These tests verify that
the errors are correctly thrown when a Q type matching these issues is
loaded.

Presently, the test for IncompatibleClassChangeError cannot be run as
support for generating objects with Q signatures is not present, just
support for Q signature fields. As a result, this test case has been
commented out.

Signed-off-by: Andrew Crowther <acrowthe3388@gmail.com>