Missing Symref for SIMD duplicated Node
In cases where a loop is a candidate for vectorization and an induction
variable is accessed in the loop, any nodes that access the induction variable
must be duplicated. There are situations where this duplicated node will not
have a correct symbol reference, and must be created after the node is
duplicated.

Signed-off-by: Ryan Santhirarajan <rsanth@ca.ibm.com>