z/TPF Compilation Failure on a2e translation wrapper
The BoolArrayStoreTransformer segment fails to compile
because of an unnecessary pre-processor redefinition
of the remove member function to atoe_remove for the
z/TPF platform only.  The redefinition occurs because
of the ascii to ebcdic translation wrapper library
required for z/TPF.  The fix is to turn on a guard
macro to prevent this translation just for this
segment.

Signed-off-by: James D Johnston <jjohnst@us.ibm.com>