Combine enter/exit tracing queries into one method
TR_J9VMBase::isMethodEnterTracingEnabled() and
TR_J9VMBase::isMethodExitTracingEnabled() are implemented as the same.
Combine them into one method TR_J9VMBase::isMethodTracingEnabled().
The tracing info sent between the JITaas client and the server is reduced
to a vector of uint8_t values instead of a vector of tuples of bool values.

Fixes: #3585

Signed-off-by: Annabelle Huo <Annabelle.Huo@ibm.com>