Improve robustness of checking tags of allocated memory blocks
J9MemTag.allocSize has type uintptr_t (aka UDATA), which is mapped
to U64 by MSVC on Windows (or with the changes in eclipse/omr#3355
on all 64-bit platforms). A random address treated as a pointer
to a J9MemTag has a 50% chance of referring to a J9MemTag where
allocSize >= 2^63 which would trigger an InvalidDataTypeException
from U64.longValue(). This change modifies ROUNDED_FOOTER_OFFSET()
so it always returns a UDATA object independent of the actual
argument type, thus avoiding that exception for blocks that otherwise
would fail the sanity checks.

Signed-off-by: Keith W. Campbell <keithc@ca.ibm.com>