Avoid mishandling of immutable objects that escape in cold blocks
In fixupNode, force non-contiguous, immutable objects that escape in
cold blocks to go through fixupFieldAccess processing to ensure the
temporary that will replace it will be properly initialized.

Then, in fixupFieldAccessForNonContiguousAllocation, ensure the store
to the original object is preserved along with the store to the temporary
that's being created to replace its field.  The former will be needed
if we escape, and the latter is needed if we don't escape.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>