Prepare NASM files to be used on Win64
Contains various minor fixes. Also, file paths need to use
backslashes on Windows. This is because NASM does not deal with
how OS file paths can differ, and simply tries searching for files
to include by prepending the include paths provided in the command
line to the file to include until it is found.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>