Add new jit hook jitFlushCompilationQueue
This new hook is needed to differenciate the case of extended HCR and
data break point. The same API jitClassesRedefined was used for both
cases before but they require different action now. Separate the code
to flush jit compilation queue into this new API which can be used for
data break point.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>