Remove redundant access mode checks in Lookup
Various access checking methods in MethodHandles.Lookup test for NO_ACCESS and
PRIVILEGED_ACCESS the same way.  In some case, the same check is performed
multiple times on the same handle.  Factor the repeated code out to a utility
method. Ensure it is called in all paths.  Rename for emphasis the methods
implementing the access checks and remove redundant checks from them.

Signed-off-by: Peter Bain <peter_bain@ca.ibm.com>