Remove unnecessary gcc-specific optimization flags
-fno-rounding-math and -fno-signaling-nans options were made
specific to GCC due to resulting in warnings on Clang. In Clang,
these options are not supported, but it does assume
round-to-nearest mode and use non-signalling NaNs by default. In
GCC, these 2 options are set by default, and hence it is not
necessary to specify them. Removing these options entirely from
the makefile will prevent any potential confusion about the
default behavior on Clang and GCC.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>