Avoid control flow for copying heap allocations to temporaries
Updates of all the temporaries that contain the address of a candidate
object for cold block heapification are currently guarded with ifacmpne
instructions.  In order to simplify the control flow graph produced for
heapification before cold blocks, replace the conditional update of each
temporary with an unconditional store using the result of an aternary
instruction that either updates the temporary with the heap object
address or with the current contents of the temporary unconditionally.
This update is inserted into the start of the cold block rather than in
a new block preceding the cold block.

This is only done if code generation supports ternary opcodes.  If
ternary opcodes are not supported, ifacmpne checks are still generated
instead.  The ifacmpne checks can also be forced by setting the
environment variable TR_disableTernaryOpForEA.

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>