Added NULL parent check to SPMDParallelizer
A problem arises if the enableGPU={enforce} Xjit option is used and
vectorization is attempted on the non-GPU path that is parallel to a newly
created JIT GPU path. Due to the enforce option, the non-GPU path is never
taken and the block is unreachable. However, the SIMD part of the
optimization still knows about the non-GPU path blocks.

This change adds a check before vectorization is done to check if the
loop to be vectorized is unreachable. It does this by checking if the
loop's parent is NULL. If it is NULL, vectorization is not performed.

Issue: #4337
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>