Compress interface dispatch tables
iTables contain vTable indices appropriate for the implementing class.
Today, iTables have a slot for every method in the interface, whether
that method has a vTable mapping or not, resulting in wasted space.

Compress iTables so they contain only appropriate methods (public
methods that appear in the vTable). For performance, overload the
totalInstanceSize field in interface J9Class structures to represent the
size of the iTable (interfaces can never be instantiated, so this field
is currently unused).

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>