Align directMethod in x86 VPIC data
Because directMethod is written at runtime, and might be concurrently
read by another thread, it must be properly aligned to ensure that
accesses to it are atomic. Currently it's possible to observe tearing,
which can lead PicBuilder to attempt direct dispatch with a bogus
J9Method pointer.

The necessary alignment is achieved by reordering the fields of the VPIC
data so that the pointer-sized fields are at the end. The VPIC data
immediately precedes a patchable instruction, which itself is aligned on
an 8-byte boundary, so this rearrangement ensures that all of the
pointer-sized fields (including directMethod) are naturally aligned.

Fixes #4446
Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>