Generate relocatable direct calls in JITaaS mode
This change is specific to the x86 codegen.

Rather than going through a snippet to perform direct
calls, we'll generate a direct call instruction to a
resolved method, which will result in a MethodCallAddress
relocation being created for the call. The relocation
record will hold the target method's client address and
when the relocation is applied on the client the direct
call will correctly point to the target.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>