Fix improper J9Class materialization during IL generation
There were a couple of instances during IL generation where the
internals of a J9Class were used to get access to a related J9Class
without using the proper APIs to do so. This was causing problems under
SVM AOT, since no validation records would be emitted when this
occurred. These have been fixed to use the proper APIs, which should
ensure that the correct validation records get emitted.

Signed-off-by: Ben Thomas <ben@benthomas.ca>