Fix object allocation inlining on Power under SVM AOT
Object allocation inlining in the Power codegen now works correctly when
the Symbol Validation Manager is enabled. The previous code was making a
number of incorrect assumptions about what would happen under AOT, which
have now been corrected.

Specifically, when allocating an array, the old AOT infrastructure would
put the address of the component class in classReg and the array class
would be retrieved from that at runtime. Under SVM AOT, it is now
possible to directly load the address of the array class, so classReg
will now contain the address of the array class and the runtime
retrieval of the array class from the component class is no longer
necessary.

Signed-off-by: Ben Thomas <ben@benthomas.ca>