A new target(java_spec_headers) is added in build-on-travis.mk
A new target is necessary because it generates jvmti header files. Since we need to generate different jvmti header files for different JDKs and building with linter doesn't depend on extension repository, we have to modify makefiles so that it generates jvmti header files when build with linter.

PRs should be commited in such order:

* jvmti.h.m4 which generates jvmti header files(#4409)
* changes in extension repositories(JDK8, JDK11, JDK12)
* changes in JDK11(#116) in its specific extension repository
* changes in how openj9 are build with linter(#4605)

There are some changes:

* A new target(java_spec_headers) is added
* A new recipe of target java_spec_headers is added: it gets java_spec_version from j9cfg.h then uses it as a variable

Signed-off-by: MarkQingGuo <Qing.Guo@ibm.com>