Disable j9process_testNewProcessGroupByTriggeringSignal on OSX
j9process_testNewProcessGroupByTriggeringSignal sends a SIGQUIT signal
to the process group. When running this test on OSX, make, tee and other
test framework processes receive a SIGQUIT signal. This causes the test
framework to crash. Due to the incompatibility with the test framework,
this test is disabled on OSX.

Issue: #4520

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>