Load correct virtualCallOffsetAndIgnoreRtResolve
There was a bug while relocating where it gets virtualCallOffsetAndIgnoreRtResolve.
Basically it was loading unsigned 32 bit integer from given pointer and casting it
to unsigned 16 bit integer which is valid virtualCallOffsetAndIgnoreRtResolve.
This leads us to load garbage instead of valid value on Big Endian platform where
required value ends up in upper half of the loaded 32 bit integer with lower half
contains garbage value. To fix this it needs to load unsigned 16 bit from the pointer.

Fixes : #4642

Signed-off-by: Rahil Shah <rahil@ca.ibm.com>