Ensure return code checked first in MXBean code
The values in Eclipse OMR for the unlimited and unknown are the same:
```
```
https://github.com/eclipse/omr/blob/cf3e1b622090ddb922717aa1b8c616ffa11c80a7/include_core/omrport.h#L161-L162

We need to differentiate between them based on the return value
of the `j9sysinfo_get_limit` call - it will return
`OMRPORT_LIMIT_UNKNOWN` in the unknown case.

Also fixed the tabs vs spaces issues in surrounding code.

Signed-off-by: Dan Heidinga <daniel_heidinga@ca.ibm.com>