Add exhaustive unit tests for toUpperCase and toLowerCase intrinsics
It seems the previous tests that were added were only testing certain
buffer lengths against certain subsets of characters. While ok, these
tests are not exhaustive and do not test the entire range of characters
which can be uppercased or lowercased.

In this commit we add exhaustive testing for both intrinsic APIs by
pre-generating a table of the uppercase and lowercase mapping. We use
this map to compare the result of calling the intrinsic.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>