Fix caseConversionHelper to check for 0xB5, 0xDF, and 0xFF corner cases
In #4597 we identified cases on which `toUpperCase` fails on Linux on
Z. This is due to some faulty login in the evaluator for these
intrinsics on Z. Bugs seem to have been introduced in #2318 for certain
characters. Namely the test in #4597 fails to convert the `\u00FF`
character. The bug is obvious, since `toUpperCase` of such a character
cannot be handled and the JIT evaluator determines the invalid
character range by checking if the character is greater than 0xFF, but
it should be checking if greater than or equal to.

In addition the compressed String evaluator also fails on 0xB5 as
determined by the unit tests. Both issues are fixed.

Fixes: #4597

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>