Clean up of Concurrent Scavenge code on Power
Comment regarding compatibility of TM and concurrent scavenge was updated.
At the current time there are no plans to be able to support both at the
time time so it was removed as being a TODO.

stopUsingRegister already performs a NULL check on the register so the
NULL checks before the calls were unnecessary.

Issue: #3421
Signed-off-by: jimmyk <jimmyk@ca.ibm.com>