Fix edge case tests to be endian aware
The edge case tests needed to store a byte into a `char[]`, but this
has to be performed in an endian aware manner because the intrinsics
act on bytes and we subsequently do an array comparison which acts on
chars. We risk the high order byte of the char being different on
little endian vs. big endian following the conversion which makes the
tests faulty.

The logical and with 0xFF00 was written with big endian in mind. Here
we make it endian aware.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>