Replace calls to indexedTrampolineLookup with findHelperTrampoline
Ask the CodeCacheManager directly for the helper trampoline address.

Change the return type of `CodeCacheManager::findHelperTrampoline` to an `intptrj_t`
and fix up references.  Introduce a J9 version of `findHelperTrampoline` to
prevent breakages while upstream OMR is changed.

Change some invocations of `findHelperTrampoline` to use the new form where the
helper index is the first parameter.

Signed-off-by: Daryl Maier <maier@ca.ibm.com>