Fix multiple problems with resolved methods
[skip ci]
- Implement version of TR_ResolvedRelocatableJ9JITaaSServerMethod::createResolvedMethodFromJ9Method
that is passed method info tuple as an argument. This version is used for most creations of resolved methods,
and we forgot to override it for AOT.

- When creating client-side mirror, need to differentiate 2 cases:
	1. Server called constructor of resolved method directly.
	2. Server called createResolvedMethodFromJ9Method.
Before this change we treated these cases as one.
It doesn't matter for non-AOT, because createResolvedMethodFromJ9Method just calls the constructor,
but it matters for AOT, because the above methods needs to do some checks, like verifying that classes
were loaded by the same class loader.

- Because of the previous point, creating a mirror can now fail.
Modify packMethodInfo to handle this case.

- Simplify constructor of TR_ResolvedRelocatableJ9JITaaSServerMethod.
We used to track on the server whether client remembered containing class of the
resolved method. This is not actually necessary, because all it affects is the
recognized method, and the correct version of that is already sent to the server.

- In TR_ResolvedRelocatableJ9Method constructor, check if owner method is not NULL,
before trying to validate class. Needed because different order of compilation object creation in JITaaS.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>