Fix code cache issues with remote AOT
[skip ci]
- Remove `prepareRelocateJITCodeAndData`, and modify
`prepareRelocateAOTCodeAndData`, to handle relocating JITaaS code.
This is done because the vast majority of the code between these
two methods is the same. The only major difference is that the former
is passed a code cache string received from the server,
while the latter assumes that code cache is contiguously allocated
right after data cache.
Address this difference by creating an optional parameter
for `prepareRelocateAOTCodeAndData`.
This also fixes a problem with remote AOT relocations,
as we were previously calling `prepareRelocateAOTCodeAndData`.

- In remote AOT mode, tell `prepareRelocateAOTCodeAndData` to allocate
code cache by passing `shouldUseCompiledCopy=false`.
In a local compilation, code cache is allocated during codegen,
so it can be later accessed when relocating code.
In JITaaS, however, we allocate code cache on the server
and need to allocate it on the client right before relocating.

- Remove assert from
`TR_JITaaSRelocationRuntime::initializeAotRuntimeInfo`
and actually implement it.
We need to set `_classReloAmount=1` for any remote compilation.

Signed-off-by: Dmitry Ten <Dmitry.Ten@ibm.com>