Add VM function to query if the JVM supports VTs
ValueTypes is an experimental feature and must be enabled with
the -XX:+EnableValhalla option. This PR also adds a J9VM helper 
for other modules to query whether valuesTypes have been enabled 
on the JDK.

Based on the changes in #4918. Adds an #ifdef guard around the 
ACC_VALUE_TYPE access bit check.

Signed-off-by: tajila <atobia@ca.ibm.com>