Use a class annotation to prevent instrumentation
New annotation: com.ibm.oti.vm.J9UnmodifiableClass

Rather than using a hard-coded (and inconsistent!) list of classes which
are not allowed to be modified by the JVMTI ClassFileLoadHook or
Re[define/transform]Classes APIs, annotate the classes which are private
to the VM implementation.

For performance, only classes loaded by the bootstrap loader will be
scanned for the annotation before ClassFileLoadHook (and that scan is
optimistic - any class file containing the bytes representing a constant
pool entry UTF8 matching the annotation signature will be tagged). This
is safe because the VM is largely in control of the boot path (it's
beyond unlikely that this pattern could match accidentally in classes
added to the boot path by the application).

Any class with access to com.ibm.oti.vm may use the annotation to
prevent being redefined. This prevents application classes from using
the annotation.

This also fixes the fact that we were not reporting ClassFileLoadHook
for any class that was not explicitly-named in the defineClass call.

Fixes: #2768

[ci skip]

Signed-off-by: Graham Chapman <graham_chapman@ca.ibm.com>