Temporarily blacklist OMRs SwitchAnalyzer.cpp
The SwitchAnalyzer optimization is being moved from OpenJ9 to OMR.
While it is being moved it will be in both repos at the same time.
For now blacklist the OMR one in OpenJ9 so that it does not find
duplicate symbols. Once a newer OMR has been accepted into OpenJ9
the OMR copy should be used and the OpenJ9 copy should be deleted.

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>