Revise tree pattern used to match inline finalizable test
The pattern used to match an inline finalizable test might be incorrect
in 64-bit mode as tree simplification might result in 32-bit operations
being used.  This change accounts for that possibility by checking
whether an l2i was applied to the result of the load of
isClassAndDepthFlags, and checks for 32-bit operations accordingly.

Fixes:  Issue #4885

Signed-off-by:  Henry Zongaro <zongaro@ca.ibm.com>