Don't copy monitor object in Read Barrier
Monitor table is weak (clearable) root and Read Barrier should not copy
objects, since info if objects is live or not is not known in a middle
of a GC cycle. The barrier only needs to update the reference slot if
object happened to be already copied.

For API with vmThread provided, also impose the assert that object is
already copied, which is true for typical synchronizing object
(vmThread->blockingEnterObject). The other API with VM provided should
be used when accessing random monitor table slots with unknown liveness
info.

Signed-off-by: Aleksandar Micic <amicic@ca.ibm.com>