Remove the unneeded if statement in ObjectMonitor.cpp::spinOnFlatLock
Referring to runtime/vm/vmthread.c#L516-L521:
In ObjectMonitor.cpp:objectMonitorEnterNonBlocking, we converted "goto
statements" into "three nested for loops". Due to this change, we can no
longer let spin counts be set to 0. With spin counts set to zero, there
will be no attempts to acquire a lock. To allow atleast one attempt to
acquire a lock, minimum value of spin counts is set to 1.

So, spinCount1/thrMaxSpins1BeforeBlocking won't be set to 0. Thus, the
unneeded if statement, (0 != spinCount1), is removed.

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>